+++ libs/local.js
@@ -102,8 +102,8 @@ module.exports = {
 
     try {
       fs.ensureDirSync(path.resolve(ROOTPATH, appconfig.paths.data))
+      fs.emptyDirSync(path.resolve(ROOTPATH, appconfig.paths.data))
       fs.ensureDirSync(path.resolve(ROOTPATH, appconfig.paths.data, './cache'))
-      fs.emptyDirSync(path.resolve(ROOTPATH, appconfig.paths.data, './cache'))
       fs.ensureDirSync(path.resolve(ROOTPATH, appconfig.paths.data, './thumbs'))
       fs.ensureDirSync(path.resolve(ROOTPATH, appconfig.paths.data, './temp-upload'))