+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Automated Upgrade / Re-install feature
+- Automated process manager
 
 ## [v1.0-beta.5] - 2017-02-12
 ### Added